+++ src/components/forms/index.ts
@@ -1,2 +1,4 @@
 export * from './Input';
+export * from './Checkbox';
 export * from './context';
+export * from './Select';