+++ stories/components/table.stories.tsx
@@ -120,3 +120,31 @@ export function HeaderBorder({ border }: TableStoryProps) {
   );
 }
 HeaderBorder.args = { border: false };
+
+export function StyledTable() {
+  return (
+    <Table style={{ color: 'white' }}>
+      <Table.Header style={{ backgroundColor: '#cccccc' }}>
+        <ValueRenderers.Title value="id" />
+        <ValueRenderers.Title value="name" />
+        <ValueRenderers.Title value="message" />
+      </Table.Header>
+
+      <Table.Row style={{ backgroundColor: 'red' }}>
+        <ValueRenderers.Text value="1" />
+        <ValueRenderers.Text value="John" />
+        <ValueRenderers.Text value="Hello" />
+      </Table.Row>
+      <Table.Row style={{ backgroundColor: 'green' }}>
+        <ValueRenderers.Text value="2" />
+        <ValueRenderers.Text value="Jane" />
+        <ValueRenderers.Text value="Hi" />
+      </Table.Row>
+      <Table.Row style={{ backgroundColor: 'blue' }}>
+        <ValueRenderers.Text value="3" />
+        <ValueRenderers.Text value="Joe" />
+        <ValueRenderers.Text value="Hey" />
+      </Table.Row>
+    </Table>
+  );
+}